metrics_pipeline/requirements.txt
requests
python-dotenv
pandas
pytest
pylint
psycopg2